feat:【数据统计】用户态数据统计- 总览 #6587

+++ src/backend/ci/core/metrics/api-metrics/src/main/kotlin/com/tencent/devops/metrics/pojo/do/AtomErrorCodeStatisticsInfoDO.kt
@@ -35,5 +35,5 @@ data class AtomErrorCodeStatisticsInfoDO(
     @ApiModelProperty("错误码信息")
     val errorCodeInfo: ErrorCodeInfoDO,
     @ApiModelProperty("错误次数")
-    val errorCount: Int
+    val errorCount: Long
 )
