+++ src/backend/ci/core/metrics/api-metrics/src/main/kotlin/com/tencent/devops/metrics/pojo/do/TurboInfoDO.kt
@@ -33,5 +33,5 @@ import io.swagger.annotations.ApiModelProperty
 @ApiModel("编译加速度量信息")
 data class TurboInfoDO(
     @ApiModelProperty("编译加速节省时间，单位：秒")
-    val turboSaveTime: Long?
+    val turboSaveTime: Double?
 )
